move aggregatexp
