+++ datafusion/physical-expr-core/src/aggregate/utils.rs
@@ -15,10 +15,23 @@
 // specific language governing permissions and limitations
 // under the License.
 
-pub mod aggregate;
-pub mod expressions;
-pub mod physical_expr;
-pub mod sort_expr;
-pub mod sort_properties;
-pub mod tree_node;
-pub mod utils;
+use std::{any::Any, sync::Arc};
+
+use super::AggregateExpr;
+
+/// Downcast a `Box<dyn AggregateExpr>` or `Arc<dyn AggregateExpr>`
+/// and return the inner trait object as [`Any`] so
+/// that it can be downcast to a specific implementation.
+///
+/// This method is used when implementing the `PartialEq<dyn Any>`
+/// for [`AggregateExpr`] aggregation expressions and allows comparing the equality
+/// between the trait objects.
+pub fn down_cast_any_ref(any: &dyn Any) -> &dyn Any {
+    if let Some(obj) = any.downcast_ref::<Arc<dyn AggregateExpr>>() {
+        obj.as_any()
+    } else if let Some(obj) = any.downcast_ref::<Box<dyn AggregateExpr>>() {
+        obj.as_any()
+    } else {
+        any
+    }
+}
